chart-service/requirements.txt
aiomysql==0.1.0
anyio==3.5.0
asgiref==3.5.0
click==8.1.2
crcmod==1.7
fastapi==0.75.1
h11==0.13.0
httptools==0.4.0
idna==3.3
pydantic==1.9.0
PyMySQL==1.0.2
python-dotenv==0.20.0
PyYAML==6.0
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.1.1
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.2
cryptography

opentelemetry-exporter-jaeger==1.11.0
opentelemetry-exporter-jaeger-proto-grpc==1.11.0
opentelemetry-exporter-jaeger-thrift==1.11.0
opentelemetry-exporter-otlp==1.11.0
opentelemetry-exporter-otlp-proto-grpc==1.11.0
opentelemetry-exporter-otlp-proto-http==1.11.0
opentelemetry-instrumentation==0.30b0
opentelemetry-instrumentation-asgi==0.30b0
opentelemetry-instrumentation-fastapi==0.30b0
opentelemetry-instrumentation-requests==0.30b0
opentelemetry-proto==1.11.0
opentelemetry-sdk==1.11.0
opentelemetry-semantic-conventions==0.30b0
opentelemetry-util-http==0.30b0
